Added Logic for EmailSend class
