+++ src/main/java/seedu/address/email/EmailSend.java
@@ -1,0 +1,96 @@
+package seedu.address.email;
+
+import java.util.Properties;
+
+import javax.mail.AuthenticationFailedException;
+import javax.mail.Message;
+import javax.mail.MessagingException;
+import javax.mail.PasswordAuthentication;
+import javax.mail.Session;
+import javax.mail.Transport;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeMessage;
+
+import seedu.address.email.exceptions.EmailLoginInvalidException;
+import seedu.address.email.exceptions.EmailMessageEmptyException;
+import seedu.address.email.exceptions.EmailRecipientsEmptyException;
+import seedu.address.email.message.MessageDraft;
+
+/**
+ * Handles how email is send via JavaAPI
+ */
+public class EmailSend {
+    private Properties props;
+
+    public EmailSend() {
+        prepEmailProperties();
+    }
+
+    /** Prepare Email Default Properties **/
+    private void prepEmailProperties() {
+        props = new Properties();
+        props.put("mail.smtp.auth", "true");
+        props.put("mail.smtp.host", "smtp.gmail.com");
+        props.put("mail.smtp.socketFactory.port", "465");
+        props.put("mail.smtp.socketFactory.class", "javax.net.ssl.SSLSocketFactory");
+        props.put("mail.smtp.port", "465");
+    }
+
+    /**
+     * Handles sending email process
+     * @throws EmailLoginInvalidException
+     * @throws EmailMessageEmptyException
+     * @throws EmailRecipientsEmptyException
+     * @throws AuthenticationFailedException
+     */
+    public void sendEmail(EmailCompose emailCompose, EmailLogin emailLogin) throws EmailLoginInvalidException,
+            EmailMessageEmptyException, EmailRecipientsEmptyException, AuthenticationFailedException {
+
+        //Step 1. Verify that the email draft consists of message and subject
+        if (!emailCompose.getMessage().containsContent()) {
+            //throw exception that user needs to enter message and subject to send email
+            throw new EmailMessageEmptyException();
+        }
+
+        //Step 2. Verify that the user is logged in with a gmail account
+        if (!emailLogin.isUserLogin()) {
+            //throw exception that user needs to enter gmail login details to send email
+            throw new EmailLoginInvalidException();
+        }
+
+        //Step 3. Verify that Recipient's list is not empty
+        if (emailCompose.getMessage().getRecipientsEmails().length <= 0) {
+            throw new EmailRecipientsEmptyException();
+        }
+
+        //Step 4. sending Email out using JavaMail API
+        sendingEmail(emailLogin.getEmailLogin(), emailLogin.getPassword(), emailCompose.getMessage());
+    }
+
+    /** Send email out using JavaMail API **/
+    private void sendingEmail(String login, String pass, MessageDraft message) throws AuthenticationFailedException {
+        final String username = login;
+        final String password = pass;
+
+        Session session = Session.getInstance(props, new javax.mail.Authenticator() {
+            @Override
+            protected PasswordAuthentication getPasswordAuthentication() {
+                return new PasswordAuthentication(username, password);
+            }
+        });
+
+        try {
+            Message newMessage = new MimeMessage(session);
+            newMessage.setFrom(new InternetAddress(username));
+            newMessage.setRecipients(Message.RecipientType.TO, message.getRecipientsEmails());
+            newMessage.setSubject(message.getSubject());
+            newMessage.setText(message.getMessage());
+
+            Transport.send(newMessage);
+        } catch (AuthenticationFailedException e) {
+            throw new AuthenticationFailedException();
+        } catch (MessagingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
